Revert "Refactor Code into Independent Files"
