+++ manifest.json
@@ -1,8 +1,8 @@
 {
 	"id": "obsidian-BibTeX-Processor",
 	"name": "BibTeX Processor",
-	"version": "1.4.0",
-	"minAppVersion": "1.0.0",
+	"version": "1.3.0",
+	"minAppVersion": "0.15.0",
 	"description": "My take on the Bibliosidian plug-in to add frontmatter to pages.",
 	"author": "rolemartyr-x",
 	"authorUrl": "https://github.com/rolemartyr-x/bibTexPlugin",
