+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "BibTeXProcessor",
-	"version": "1.4.0",
+	"version": "1.3.0",
 	"description": "My take on the Bibliosidian plug-in to add frontmatter to pages.",
 	"main": "main.js",
 	"scripts": {
