+++ versions.json
@@ -2,6 +2,5 @@
 	"1.0.0": "0.15.0",
 	"1.1.0": "0.15.0",
 	"1.2.0": "0.15.0",
-	"1.3.0": "0.15.0",
-	"1.4.0": "1.0.0"
+	"1.3.0": "0.15.0"
 }
